test: Added .env toggles for local testing and screenshot directory
