+++ .env.example
@@ -1,0 +1,7 @@
+# Test environment configuration
+
+# Set 'local' to run tests on the local machine, 'docker' (default) to run tests in a Docker container.
+TEST_ENV=docker
+
+# Directory path to save screenshots of tests. No screenshots will be saved if the variable is not set. 
+SNAPSHOT_SAVE_DIR=./playwright-screenshots
