+++ .gitignore
@@ -15,3 +15,5 @@
 /test-results
 /playwright-report
 /graphql-voyager-*.tgz
+/playwright-screenshots
+.env
