+++ package.json
@@ -77,6 +77,7 @@
     "@typescript-eslint/parser": "8.16.0",
     "cspell": "9.4.0",
     "css-loader": "7.1.2",
+    "dotenv": "^17.3.1",
     "eslint": "8.57.1",
     "eslint-plugin-import": "2.31.0",
     "eslint-plugin-playwright": "1.6.2",
