+++ tests/pageObjectModel/index.ts
@@ -1,0 +1,6 @@
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+export { gotoVoyagerPage } from './goToVoyagerPage.ts';
+export { SchemaPresets } from './schema/index.ts';
