+++ tests/pageObjectModel/schema/index.ts
@@ -1,0 +1,3 @@
+export * from './changeDialog.ts';
+export * from './presets.ts';
+export * from './tabs.ts';
