+++ tests/pageObjectModel/schema/presets.ts
@@ -1,0 +1,6 @@
+export const SchemaPresets = [
+  'Star Wars',
+  'Yelp',
+  'Shopify Storefront',
+  'GitHub',
+] as const;
